+++ Package.swift
@@ -7,7 +7,7 @@
 
 import PackageDescription
 
-let version = "1.10.1"
+let version = "1.10.2"
 let packageLibraryName = "Compatibility"
 
 // Products define the executables and libraries a package produces, making them visible to other packages.